Add files via upload
John Kenny & Sons Contractor as a company provided raw data on jobs that have been awarded to them over a year. 
From the data, they would like to find out:

i. the split between the jobs that were ‘completed’ and the jobs that were  ‘cancelled’ 
ii. in form of a table, the break down of ‘job type’ including both the count and % of overall jobs
iii. the percentage of jobs were completed within their target date
iv. in form of another table, the average time taken to complete jobs by their trade type
v. the sector(s) are jobs taking longest on average to complete